requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
deep-translator==1.11.4
frozenlist==1.4.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
YooMoney==0.1.0
